+++ hss_okta_flutter/lib/src/web/stubs/hss_okta_flutter_web_stub.dart
@@ -63,4 +63,6 @@ class HssOktaFlutterWeb {
 
   Future<bool> hasTokenExpired(AbstractToken token) async =>
       throw UnimplementedError();
+
+  Future<UserClaims> getUser() => throw UnimplementedError();
 }